taulu-api/pyproject.toml
[project]
name = "taulu-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
